split S3 upload from download
