+++ spec/services/storage/s3/downloader_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Storage::S3 do
+RSpec.describe Storage::S3::Downloader do
   let(:test_client) { Aws::S3::Client.new(stub_responses: true) }
 
   before :each do
@@ -15,34 +15,18 @@ RSpec.describe Storage::S3 do
     described_class.new(path: path)
   end
 
-  describe '#upload' do
-    before :each do
-      test_client.stub_responses(:head_object, [
-        false,
-        { content_length: 150 }
-      ])
-    end
-
-    it 'uploads file to s3' do
-      expect(subject.exists?).to be_falsey
-      subject.upload
-      expect(subject.exists?).to be_truthy
-    end
-
-    after :each do
-      subject.purge_from_s3!
-    end
-  end
-
   describe '#download' do
+    let(:uploader) { Storage::S3::Uploader.new(path: path) }
+
     before :each do
+      uploader.upload
+
       test_client.stub_responses(:get_object, [
         { body: "lorem ipsum\n" }
       ])
     end
 
     it 'downloads file from s3' do
-      subject.upload
       subject.download
 
       downloaded_path = subject.send(:temp_file).path
